# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Saashakti API",
    "version": "0.1.0",
    "description": "Saashakti backend — OTP onboarding, Aadhaar (mock), beneficiary, scheme matching, dashboard."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "AadhaarStatus": {
        "type": "string",
        "enum": [
          "not_started",
          "pending",
          "verified",
          "failed",
          "mock_verified"
        ]
      },
...
    }
  },
  "paths": {
...
    "/v1/aadhaar/start": {
      "post": {
        "summary": "Start Aadhaar verification (mock provider)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "$ref": "#/components/schemas/AadhaarStatus"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/aadhaar/status": {
      "get": {
        "summary": "Get Aadhaar verification status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Status"
          }
        }
      }
    },
    "/v1/aadhaar/kyc": {
      "get": {
        "summary": "Mock e-KYC fetch (returns name, DOB, gender, address, mobile from Aadhaar)",
        "parameters": [
          {
            "name": "aadhaarNumber",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 12,
              "maxLength": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "KYC record"
          },
          "400": {
            "description": "Invalid Aadhaar number"
          }
        }
      }
    },
...
  }
}